adding IO tweaks for converting to multiple steps per day

+++ src/main/java/uk/co/ramp/io/types/CmptRecord.java
@@ -29,7 +29,7 @@ public interface CmptRecord {
         .build();
   }
 
-  int time();
+  double time();
 
   int s();
